sys/src/9/pcboot: use uintptr for the boot malloc

+++ sys/src/9/pcboot/expand.c
@@ -223,7 +223,7 @@ isgzip(uchar *p)
 }
 
 void*
-malloc(ulong n)
+malloc(uintptr n)
 {
 	void *v;
